Refactor return type of @operator decorator
Having the operator decorator return a type, whilst accurate,
was causing issues with the parameters to __init__ being confused.

Instead, a Protocol is defined with a __call__ method returning
an instance of (a subclass of the) Stream class, along with a
raw method retaining the original signature.
This is, from a logic standpoint, the same API, just masquerading
the creation of a dynamic type with an instance of a known type.

This also allows us to properly deliniate between the return types
when pipable is set -- we introduce a PipeOperator that extends
the StreamOperator by adding the pipe method.